refactor: split visual effects into separate files
